Add support for ai players
