+++ sparkle-phoenix.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const uuid = require('uuid');
 const HttpServer = require('./src/httpServer');
 const WSManager = require('./src/wsManager');
 const Game = require('./src/game');
@@ -9,9 +10,12 @@ const AiPlayer = require('./src/aiPlayer');
 const simMode = Common.getArgument('sim', false);
 const observerMode = Common.getArgument('observe', false);
 const playerCount = Common.getArgument('players', 2);
+const playerCountAI = Common.getArgument('ai-players', 0);
 const simModeTickSpeed = Common.getArgument('tick', 2000);
 const helpMode = Common.getArgument('help', false);
 const shrapnelMode = Common.getArgument('no-shrapnel', false);
+
+const aiPlayer = new AiPlayer();
 
 if (helpMode) {
   console.log(`
@@ -19,6 +23,7 @@ if (helpMode) {
 
   args:
   -players [number]     Player count
+  -ai-players [number]  Number of AI Players
   -sim                  Simulation mode
   -observe              Wait for an observer to join before starting Simulation mode
   -tick [number]        Milliseconds to wait between each making each move during Simulation mode
@@ -30,7 +35,7 @@ if (helpMode) {
 if (simMode) {
   var wsManager = new WSManager();
   var httpServer = new HttpServer(80, wsManager);
-  var sim = new Sim(simModeTickSpeed, new AiPlayer());
+  var sim = new Sim(simModeTickSpeed, aiPlayer);
 
   httpServer.startServer();
 
@@ -61,9 +66,16 @@ if (simMode) {
 
   httpServer.startServer();
 
+  var aiPlayers = [];
+
   var game = new Game(function(id, state) {
     console.log(`Callback = ID: ${id}`);
     console.log(state);
+    if (aiPlayers.includes(id)) {
+      console.log('Making move for AI player');
+      let newMove = aiPlayer.determineBestMove(game, id);
+      game.endTurn(id, newMove[0], newMove[1]);
+    }
     wsManager.sendMessage(id, state);
   }, function(){process.exit()}, shrapnelMode);
 
@@ -81,22 +93,19 @@ if (simMode) {
   })
   
   game.newGame(playerCount);
+  
+  generateAIPlayers(game, aiPlayers);
 }
 
-
-
-
-
-/*
-
-app.delete('/logout', function (request, response) {
-  const ws = map.get(request.session.userId);
-
-  console.log('Destroying session');
-  request.session.destroy(function () {
-    if (ws) ws.close();
-
-    response.send({ result: 'OK', message: 'Session destroyed' });
-  });
-});
-*/
+function generateAIPlayers(game, aiPlayers) {
+  let newId;
+  while (aiPlayers.length < playerCountAI) {
+    newId = uuid.v4();
+    
+    if (!aiPlayers.includes(newId)) {
+      console.log('Adding a new AI player with ID: ' + newId);
+      aiPlayers.push(newId);
+      game.playerJoin(newId);
+    }
+  }
+}
